requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
jose
bcrypt
slowapi
loguru
